+++ packages/gateway/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @ai-sdk/gateway
 
+## 4.0.55
+
+### Patch Changes
+
+- 1f7a464: chore(provider/gateway): update gateway model settings files
+
+## 4.0.54
+
+### Patch Changes
+
+- 257632b: chore(provider/gateway): update gateway model settings files
+
+## 4.0.53
+
+### Patch Changes
+
+- a0b1ffc: chore(provider/gateway): update gateway model settings files
+
 ## 4.0.52
 
 ### Patch Changes